backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-CORS==3.0.10
Flask-Limiter==3.3.0
python-dotenv==0.19.0
requests==2.26.0
beautifulsoup4==4.10.0
APScheduler==3.8.0
gunicorn==20.1.0
pytest==6.2.5
psycopg2-binary==2.9.3

asyncio==3.4.3
